requirements.txt
cryptography==41.0.4
pynput==1.7.6
python-dotenv==1.0.0